backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # Commented out - use SQLite for MVP
alembic==1.12.1

# Vector DB and Embeddings
# Using pure Python implementation - no compilation needed
sentence-transformers==2.2.2
# We'll use a simple in-memory vector store for MVP

# LLM and AI
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1
google-generativeai==0.3.2

# PDF Processing
PyPDF2==3.0.1
python-magic==0.4.27

# Utilities
python-dotenv==1.0.0
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.25.2

# Authentication (optional for MVP)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS is built into FastAPI, no additional package needed
